Add Region object iterator and clean up FreeList
The region object iterator works by walking a valid markmap.  There is
an adapter for easily using the iterator in range based for loops.

The free list has been modified to keep a pointer to the last block
added.  This allowed FreeBlocks to be stored in the FreeList in the
order that they are added.

Signed-off-by: Andrew Young <youngar17@gmail.com>