Disable static analysis tools and remove bogus header include (#2)
Add an OmOption utility and use it

This is a utility that helps define public boolean flags. Default values treated as if-expressions. Requires and Conflicts are lists of flag names that must be true/false depending on the value of the flag being defined.

Option groups have to be explicitly validated after all options are defined.