Rearrange CMake project component options
This renames each of the component names to begin with `OM_COMP`, which
aligns with the project rebranding to `OM`.  This also properly creates
the dependencies between the different projects and sets up the caches
to rename them.  Some component options were renamed to reflect the
specific component they belong to.  All options are still in the top
level directory to enable option verification/validation before we start
processing subdirectories, which may depend on the options being
correct.