Swap the order of the visitor and context arguments for walk
The walk method should take a visitor, plus a variadic tail of arguments
to forward along. The context is one such "forwarded" argument.

Giving the context "pride of place" at the front of the argument list
makes the visitation/walk API less general. If I want to just print the
slots in an object, using a functor, why do I need to provide a context?